python-service/requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0

# Machine Learning
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.2

# Time series specific
statsmodels==0.14.1
prophet==1.1.5

# Utilities
joblib==1.3.2
matplotlib==3.8.2
